Migrate transfer functions to header to get these templated functions working
